+++ src/react/components/checkbox/XUICheckbox.js
@@ -124,7 +124,7 @@ const buildCheckbox = (qaHook, htmlClassName, svgSettings, isGrouped) => {
  * @extends {Component}
  */
 export default class XUICheckbox extends Component {
-	// User can manually proivde an id, or we will generate one.
+	// User can manually provide an id, or we will generate one.
 	wrapperIds = generateIds(this.props.labelId);
 
 	componentDidMount() {
@@ -304,7 +304,7 @@ XUICheckbox.propTypes = {
 	/** The tab-index property to place on the checkbox */
 	tabIndex: PropTypes.number,
 
-	/** Used to output an uncontrolled checkbox component.  If a value is passed to the
+	/** Used to output an uncontrolled checkbox component. If a value is passed to the
 	 * isChecked prop, this prop will be ignored. */
 	isDefaultChecked: PropTypes.bool,